Introduce custom property mechanism
This is rather elaborate, and very handy.

Create a file in ./properties and use it to define a custom
target property, and also an associatd INTERFACE target. This can
be used to rather handily group the effects of a particular
feature together into a handy unit, that is then automatically
applied if the corresponding property is set on a target.

See the comment in Targets.cmake for a less terrible explanation.

Used to implement sanitiser and vectoriser remarks properties,
and readily extensible to more...