Add a mechanism for optional compiler flags
Many compiler flags, such as diagnostics or optimiation flags,
aren't really mandatory. It's annoying to put in version checks
for them, and it's annoying to make xcmake only work with specific
compiler versions, so instead let's use cmake's existing system
for checking if the compiler accepts a certain flag. Now we can
use `target_compile_options()` for mandatory flags, and this
new function for optional flags (which will simply omit any that
are not supported).