Use `${CMAKE_COMMAND}` to avoid an obsure problem
`CMAKE_COMMAND` gives you the full path to the same cmake binary
you are running inside. This is most certainly the cake you want
to use for `cmake -E`, or all kinds of strange things can happen.

Making a custom command that just calls `cmake` is relying on the
platform-specific `PATH` resolution to find a cmake to use. That
might give you a different cmake, or even no cmake at all (causing
our build system to fail on platforms that don't have cmake in
`PATH` but which invoke it explicitly anyway).