Extend the trademark compliance to the header precompilation step
This also sets the stage for partial preprocessing of headers:
something we've planned to do for a while for various reasons
but haven't yet bothered with.

The purpose of this is that headers produce Doxygen documentation,
and it should comply. Analysis of Doxygen documentation will
produce errors that are tiresome to map to header source
locations, so we may as well just analyse all the headers (which,
presumably, also need to be compliant).