Always pass `--toc` to Pandoc
Pandoc and tables of contents are annoying.

You can set whether or not a TOC is generated for a given
document using a metadata block at the top:

```
---
toc: false
---
```

However, if `--toc` is not passed to the `pandoc` invocation then
setting `toc:true` will not lead to one being generatred. The
only way to do it is to pass `--toc` always and then set
`toc:false` on pages you don't want a table of contents.

So let's do that.