Handle the insane inconistency with absolute install paths
So, the newer version of cmake treats an absolute install
destination as a true absolute install destination, without
prefixing the install prefix.

On Windows, however, it does not (because `/` is not considered
an absolute path there). Hence why earlier work in this area
made everything burst into flames, but only on Linux.

Fun fun fun.