Use cmake-powered sed metaprogramming to fix documentation URLs
The goal here is to make documentation hyperlinks always work.
- When viewed on gitlab
- When viewed in an install tree
- When viewed from an installed package (which has a different
  directory layout to a classic install tree).

After this commit, documentation hyperlinks should be written as
if you're targeting gitlab. That is:
- Links to other markdon files in this repo can be relative or
  absolute URLs to the file on gitlab.
- Links to markdown file in other repos should be absolute links
  to that file on gitlab (not an attempt to guess the relative
  path in the output install tree, because this is prone to
  change)
- Links to Doxygen pages should link to the corresponding page on
  the VPN.

This convention means such links are canonicalised, and the build
system can relatively easy rewrite all of these to point to where
they need to go.