Make the GPU autodetection mechanism handle multiple architectures
This causes the default native-mode to compile a fatbinary for
every GPU found in the build machine.

Being explicit is likely to be what people want if they don't
want builds to take an insanely long time.

The speclib assumption cache has not yet been updated, so
multi-gpu builds may still be reporting incorrect device properties.