bugfix: Make Doxygen regenerate on header changes again.
Seems to be introduced in 13863db27000dad1dbee37b1e8e950ebfdaacca1.
The doxygen target code expected NEW_PATHS to be a list of directories
that it would add all headers from, while after that commit NEW_PATHS
became a list of headers.

During debugging I found out that the tagfile for the target is not
actually mentioned in the dependencies of the target
(CMakeFiles/TARGET.dir/all in build/CMakeFiles/Makefile2) but it
still works fine so I guess I just don't know how CMake does it.