Fix the printing of warnings at the end
This got removed as part of another change. But also, someone ran
into problems with an infinite loop on Windows. Actually, that's
caused by a cmake bug, I think, that got fixed in version 3.19.
So partly because I'm not sure the partial removal of the feature
was intended, and partly because I think the problem it caused by
a bug, I'm putting the feature back.