Make target_link_libraries use a keyword argument to replicate the default after emitting a warning
Prior, it would call the original target_link_libraries, but
without the keyword argument. This makes it incompatible with
add_library and friends, which use the keyword version. The
documentation says the keywordless version behaves in the way a
PUBLIC keyword would.

This was added to enable building third party projects such as
zlib.