Don't static-link the standard library when msan is in use
Crucially, `-static-libstdc++` disables libc++.

At present the only way to static-lib libc++ is to tell the compiler
not to implicitly link the standard library at all, and then
manually reintroduce it with a pile of other flags. This seems
not worth the effort, since there's an upstream issue to add
an equivalent flag for static-libc++, and the non-static-linked
case works just fine for msan runs anyway.