Exclude headers from find_sources
We did this years ago because it was necessary to make the IDE index
header files. It would regard them as not being part of the project.

These days, clion got smarter so this is not necessary, and it's
getting in the way. We could have cmake scripts that run aronud
and se HEADER_FILE_ONLY on everything, but that'd be kinda slow.