Don't test cuda flags without a CUDA compiler
The check_cuda_compiler_flag command greps for phrases which appear when
NVCC (called via execute_process) is passed an unrecognised flag. When
CMAKE_CUDA_COMPILER is not set, the output is in reference to an invalid
shell command, not NVCC diagnostics, so the test passes.

If CUDA is enabled after xcmake initialization, the flags which were
incorrectly tested will be applied to CUDA targets.

While the previous behaviour was not necessarily *wrong* (it simply
assumes that CMAKE_CUDA_COMPILER is set), it is unintuitive both in
implementation and in its result.

Furthermore, perhaps the user would like to have xcmake features
available to them as they do their own tests for CUDA support. In which
case it *is wrong* for xcmake to assume that if it doesn't have CUDA,
then the project will never use CUDA.