Don't test cuda flags without a CUDA compiler
The check_cuda_compiler_flag command greps for phrases which appear when
NVCC (called via execute_process) is passed an unrecognised flag. When
CMAKE_CUDA_COMPILER is not set, the output is in reference to an invalid
shell command, not NVCC diagnostics, so the test passes.

If CUDA is enabled after xcmake initialization, the flags which were
incorrectly tested will be applied to CUDA targets.