Remove fast-math from the xcmake OPT_LEVEL stuff
With the retirement of `-Ofast`, and the introduction of far
more varied fast-math flags than once existed, this probably
needs to be handled specifically by projects.

A notable pitfall of the old approach was that enabling
fast-math flags on some TUs but not others would not work
as desired, since some of the new fast-math stuff is actually
based on link-time stuff, so affects an entire library/executable
regardless.