Don't find SCALE package inside scale tree
An export(EXPORT) call does not complete until configuration is
complete, so a scale package is not usable until configuration is done,
so find_package from a submodule won't work.

Fix is to just check if we are in the scale build tree. If so, we can
just link targets directly without needing to get the package.