Add back search functions
These were removed due to being broken when passed NO_CACHE, as OUTVAR
would be treated as a regular variable which would be set at the scope
of the function override, not the caller.

Now OUTVAR is correctly set with PARENT_SCOPE if NO_CACHE is present,
and `mark_as_advanced`'d otherwise.